boards: xtensa: support run the MTL simulator out of tree [INTERNAL]
Add support for running the mtl simulator from specifying
directory. It will run the MTL simulator in "$HOME/mtlsim"
by default, if the evironment variable "MTL_SIM_DIR" is
not specified.

That means we need to put the simulator related file into
this directory. There are several files are needed:

1. The dsp_fw_sim, the simulator binary.
2. The dsp_rom_mtl_sim.hex, the ROM file.
3. The other libraries that simulator is needed. These
   libraries depends on which version of MTL simulator
   you are going to use.

And these files will be provided in other places.

Signed-off-by: Enjia Mai <enjia.mai@intel.com>