intel_adsp: ace: Fix z_mapped_end in the linker [INTERNAL]
z_mapped_end was set in the uncached region while z_mapped_start
was inside the cached region. This end up with z_mapped_end in a lower
address than z_mapped_start.

Since the kernel is mapped in the cached region, lets map this region.
Another reason to do this is that if we use these variables to map
uncached and cached region we will end up mapping memory that is
not used since these regions are not contiguous.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>