xtensa: mmu: disable array bound warnings on SoC MMU ranges
GCC complains about usage of the SoC MMU range array
(xtensa_soc_mmu_ranges) as the default weak declaration is
an empty array, and any access to its element is considered
out of bound access. However, we have a number of element
variable to guard against this (... if done correctly).
Besides, this will almost be overridden by the SoC layer.
So tell GCC to ignore this.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>