twister: add support for plans and test configuration
Add support test plans and the ability to assign a specific test to one
or more plans. Use command line options of twister it is then possible
to select a plan and just execute the tests included in this plan.

Additionally, a test configuration is now available to define plan
dependencies and additional inclusion of tests into a specific plan if
the test itself does not have this information already.

In the configuration file you can include complete components using
regular expressions and you can specific which test plan to import from
the same file, making management of plans easier.

To help with testing outside of upstream CI infrastructure, additional
options are available in the configuration file, which can be hosted
locally. As of now, those options are available:

- Ability to ignore default platforms as defined in board definitions
  (Those are mostly emulation platforms used to run tests in upstream
  CI)
- Option to specify your own list of default platforms overriding what
  upstream defines.
- Ability to override build_onl_all options used in some testscases.
  This will treat tests or sample as any other just build for default
  platforms you specify in the configuation file or on the command line.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>